🐛 Use web/app src for local & web/dist for build/docker
The main, but smallest change for this is in `cmd/mcp-standalone/main.go` to change
the default value of `-web-dir` to `./web/app`.
This allows serving of static assets for the UI for local development
without extra config needed.

The bulk of this change extends the `make image` task to build the web
assets and include them in the docker image.
The OpenShift deployment config is updated to specify `-web-dir /web/dist`.
All OpenShift resources are updated to use `mcp-standalone` instead of
`mobile-server` for naming of resources & the server binary.

This also fixes #12 by using a redirectreference for oauth client to the
mcp-standalone route.
A local dev version, `sa.local.json`, can be used when running the server
outside of OpenShift. It uses a hardcoded redirecturi of
`https://127.0.0.1:3001`.

The Oauth flow in the UI will not work yet in OpenShift until #9, #10 &