🐳 Use paths under ./ui for OpenShift data dirs
This change configures `oc cluster up` to use dirs under the 'ui'
folder, similar to the `--host-config-dir`, for below

- `--host-data-dir` ./ui/openshift-data
- `--host-pv-dir` ./ui/openshift-pvs
- `--host-volumes-dir` ./ui/openshift-volumes

This allows for 2 things:
- easier cleanup of an oc cluster by running the clean.sh script
(Removes PV data too)
- data storage for PV/Volumes/Data is kept where the repo is cloned.
-- this change helps with a standard enough linux setup where the docker
storage may be on the same mount point as the default location for these
`oc cluster` things. This could lead to low disk space after using
docker & openshift for a while. Using the setup in this commit will
allow more space for docker stuff, and keep the openshift stuff separate