🎨 Move the external service modal out of the main content div
This change prevents classes/styles on the main content div from
affecting thing modal.
The knock on effect of this is the modal launch button is no longer part
of the model itself. It is expected to launch the modal whatever way
you want rather than have the modal directive automatically insert a
launcher button.