Setup admin user with cluster edit permissions
Currently we have to create this user manually. This change adds a
role that will create a user named admin and give it cluster edit
permissions.

A further change has been made to provide additional labels in the
playbook file to differentiate between what should be done for
normal local installations and local dev installations.

When running the playbook with --tags=local the admin user will not
be created. It will with local-dev.